requirements.txt
absl-py==1.0.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
autopep8==1.6.0
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
dj-database-url==0.5.0
Django==3.2.11
executing==0.8.3
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
ipython==8.3.0
jedi==0.18.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
protobuf==3.20.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
traitlets==5.1.1
typing_extensions==4.2.0
Unipath==1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
whitenoise==5.3.0
wrapt==1.14.1
